Add files via upload
This is the first step to a successful development of GEE development which involves the proper connection to GEE.  Most people could not successfully connected to the GEE and ger stuck in the processing of ee.Initialize(). As shown, we need first to import os as an additional library to connect the server of Goolge Earth Engine.